+++ Paradigm/src/utopia/paradigm/shape/shape2d/area/polygon/Triangle.scala
@@ -1,15 +1,19 @@
 package utopia.paradigm.shape.shape2d.area.polygon
 
-import utopia.flow.collection.immutable.Single
+import utopia.flow.collection.immutable.{Pair, Single}
+import utopia.flow.operator.combine.Combinable
 import utopia.flow.util.Mutate
 import utopia.paradigm.shape.shape2d.Matrix2D
 import utopia.paradigm.shape.shape2d.vector.Vector2D
 import utopia.paradigm.shape.shape2d.vector.point.Point
 import utopia.paradigm.shape.shape3d.Matrix3D
+import utopia.paradigm.shape.template.HasDimensions.HasDoubleDimensions
 import utopia.paradigm.transform.Transformable
 
 object Triangle
 {
+	// OTHER    ------------------------------
+	
 	/**
 	  * Creates a new triangle with three corners provided
 	  * @param p1 The first corner
@@ -17,24 +21,50 @@ object Triangle
 	  * @param p3 The third corner
 	  * @return A new triangle
 	  */
-	def withCorners(p1: Point, p2: Point, p3: Point) = new Triangle(p1, (p2 - p1).toVector, (p3 - p1).toVector)
+	def withCorners(p1: Point, p2: Point, p3: Point): Triangle = TriangleWithCorners(Vector(p1, p2, p3))
+	
+	/**
+	  * @param origin Primary corner of this triangle
+	  * @param sides Two sides that leave from this corner, as vectors
+	  * @return A new triangle
+	  */
+	def apply(origin: Point, sides: Pair[Vector2D]): Triangle = TriangleWithSides(origin, sides)
+	/**
+	  * @param origin Primary corner of this triangle
+	  * @param side1 The first side that leaves from this corner, as a vector
+	  * @param side2 The second side that leaves from this corner, as a vector
+	  * @return A new triangle
+	  */
+	def apply(origin: Point, side1: Vector2D, side2: Vector2D): Triangle = apply(origin, Pair(side1, side2))
+	
+	
+	// NESTED   -------------------------------
+	
+	private case class TriangleWithSides(origin: Point, primarySides: Pair[Vector2D]) extends Triangle
+	{
+		// ATTRIBUTES   ----------------
+		
+		lazy val corners = origin +: primarySides.map { origin + _ }
+		
+		
+		// IMPLEMENTED  ----------------
+		
+		override def edges =
+			Vector(primarySides.first, primarySides.reverseReduce { _ - _ }, primarySides.second)
+		
+		override def +(other: HasDoubleDimensions) = copy(origin = origin + other)
+	}
+	
+	private case class TriangleWithCorners(corners: Seq[Point]) extends Triangle
 }
 
 /**
-  * This shape represents a triangle
-  * @param origin The origin point of the triangle
-  * @param side1 The first side of this triangle as a vector
-  * @param side2 The second side of this triangle as a vector
+  * Common trait for polygons with 3 corners.
+  * @author Mikko Hilpinen
+  * @since ???, < v1.7
   */
-// TODO: Change into a trait and allow 2 formats: sides or corners
-case class Triangle(origin: Point, side1: Vector2D, side2: Vector2D)
-	extends Polygonic with Transformable[Triangle]
+trait Triangle extends Polygon with Transformable[Triangle] with Combinable[HasDoubleDimensions, Triangle]
 {
-	// ATTRIBUTES   ----------------
-	
-	lazy val corners = Vector(origin, origin + side1, origin + side2)
-	
-	
 	// IMPLEMENTED	----------------
 	
 	override def identity = this
@@ -53,14 +83,14 @@ case class Triangle(origin: Point, side1: Vector2D, side2: Vector2D)
 		math.sqrt(s * sideLengths.view.map { s - _ }.product)
 	}
 	
-	override def edges = Vector(side1, side2 - side1, side2)
-	
 	override def toTriangles = Single(this)
 	override def convexParts = Single(this)
 	override def collisionAxes = edges.map { _.normal2D }
 	
 	override def transformedWith(transformation: Matrix3D) = map { _ * transformation }
 	override def transformedWith(transformation: Matrix2D) = map { _ * transformation }
+	
+	override def +(other: HasDoubleDimensions) = map { _ + other }
 	
 	override def map(f: Mutate[Point]) = {
 		val mapped = corners.map(f)
@@ -71,9 +101,9 @@ case class Triangle(origin: Point, side1: Vector2D, side2: Vector2D)
 	// OTHER    -------------------
 	
 	/**
-	 * @param f A mapping function
-	 * @return A copy of this triangle with mapped corners
-	 */
+	  * @param f A mapping function
+	  * @return A copy of this triangle with mapped corners
+	  */
 	@deprecated("Renamed to .map(...)", "v1.7")
 	def mapCorners(f: Point => Point) = map(f)
 }