+++ Paradigm/src/utopia/paradigm/shape/shape2d/area/polygon/c4/package.scala
@@ -1,0 +1,14 @@
+package utopia.paradigm.shape.shape2d.area.polygon
+
+/**
+  * @author Mikko Hilpinen
+  * @since 25.08.2024, v1.7
+  */
+package object c4
+{
+	/**
+	  * An alias for Parallelogram. For backwards-compatibility.
+	  */
+	@deprecated("Please use Parallelogram instead", "v1.7")
+	type Parallelogramic = Parallelogram
+}