+++ Paradigm/src/utopia/paradigm/shape/shape2d/area/polygon/package.scala
@@ -1,0 +1,14 @@
+package utopia.paradigm.shape.shape2d.area
+
+/**
+  * @author Mikko Hilpinen
+  * @since 25.08.2024, v1.7
+  */
+package object polygon
+{
+	/**
+	  * An alternative name for the Polygon trait, for backwards-compatibility
+	  */
+	@deprecated("Please use Polygon instead", "v1.7")
+	type Polygonic = Polygon
+}